% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow_generics.R
\name{download_file}
\alias{download_file}
\title{download an arbitrary file from S3}
\usage{
download_file(key, s3, dest = basename(key))
}
\arguments{
\item{key}{the file name (key) to be used in the bucket (string, can include path)}

\item{s3}{an arrow FileSystem object (usually S3 Filesystem)}

\item{dest}{file name or connection where file should be written locally}
}
\description{
Note: file must be small enough to fit in RAM with this method.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
s3 <- arrow::s3_bucket("neon4cast-forecasts", 
                       endpoint_override = "data.ecoforecast.org", 
                       anonymous=TRUE)
example <- tail(s3$ls("phenology"), 1)
dest <- tempfile()
download_file(example, s3, dest)
\dontshow{\}) # examplesIf}
}
