% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow_generics.R
\name{put_object}
\alias{put_object}
\title{upload an object to an arrow filesystem (such as remote S3 bucket)}
\usage{
put_object(obj, key, s3, fun = readr::write_csv, ...)
}
\arguments{
\item{obj}{an R object that can be serialized to disk by \code{fun} (e.g. data.frame)}

\item{key}{the file name (key) to be used in the bucket (string, can include path)}

\item{s3}{an arrow FileSystem object (usually S3 Filesystem, see arrow::s3_bucket)}

\item{fun}{a function that can serialize obj (to an anonymous file).}

\item{...}{additional arguments to \code{fun}}
}
\description{
upload an object to an arrow filesystem (such as remote S3 bucket)
}
