% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_forecast.R
\name{read_forecast}
\alias{read_forecast}
\title{read a forecast file (csv or netcdf)}
\usage{
read_forecast(
  file_in,
  target_variables = c("oxygen", "temperature", "chla", "richness", "abundance", "nee",
    "le", "vswc", "gcc_90", "ixodes_scapularis", "amblyomma_americanum"),
  reps_col = "ensemble",
  quiet = TRUE,
  ...
)
}
\arguments{
\item{file_in}{path or URL to forecast to be read in}

\item{target_variables}{vector of valid variables being predicted}

\item{reps_col}{name of the replicates column (for ensemble forecasts)}

\item{quiet}{logical, default TRUE (show download progress?)}

\item{...}{additional options (currently ignored)
Reads in a valid forecast as a data.frame() in EFI format.
csv files are simply read in using \code{readr::read_csv}.
This utility  exists mostly to provide a single function that can
handle both csv and netcdf formats, as both are valid choices in EFI standard.}
}
\description{
read a forecast file (csv or netcdf)
}
