% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow_generics.R
\name{upload_file}
\alias{upload_file}
\title{upload a file from local disk to an S3 bucket}
\usage{
upload_file(file, key = file, s3)
}
\arguments{
\item{file}{file name or connection of file to be uploaded}

\item{key}{the file name (key) to be used in the bucket (string, can include path)}

\item{s3}{an arrow FileSystem object (usually S3 Filesystem)}
}
\description{
Note: file must be small enough to fit in RAM with this method.
}
